The Agentic Toggle
Single agentic switch to toggle the AssistantCoordinator from the old behavior to the new behavior.

Migrate SSE helper functions and their tests from gemini to sse.
